Using Automapper

+++ SchedulerApi/Controllers/ScheduleApiController.cs
@@ -1,4 +1,6 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using AutoMapper;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
 using SchedulerApi.Data;
 using SchedulerApi.Model;
 using SchedulerApi.Model.dto;
@@ -10,29 +12,31 @@ namespace SchedulerApi.Controllers
     public class ScheduleApiController : ControllerBase
     {
         private readonly ApplicationDbContext _db;
-        public ScheduleApiController(ApplicationDbContext db)
+        private readonly IMapper _mapper;
+        public ScheduleApiController(ApplicationDbContext db, IMapper mapper)
         {
             _db = db;
+            _mapper = mapper;
         }
 
         [HttpGet]
         [ProducesResponseType(StatusCodes.Status200OK)]
-        public ActionResult<IEnumerable<ScheduleDTO>> GetSchedules()
+        public async Task< ActionResult<IEnumerable<ScheduleDTO>>> GetSchedules()
         {
-            return Ok(_db.Schedules.ToList());
+            return Ok(await _db.Schedules.ToListAsync());
         }
 
         [HttpGet("{id:int}", Name = "GetSchedule")]
         [ProducesResponseType(StatusCodes.Status200OK)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
         [ProducesResponseType(StatusCodes.Status404NotFound)]
-        public ActionResult<ScheduleDTO> GetSchedule(int id)
+        public async Task< ActionResult<ScheduleDTO>> GetSchedule(int id)
         {
             if (id == 0)
             {
                 return BadRequest();
             }
-            var schedule = _db.Schedules.FirstOrDefault(u => u.Id == id);
+            var schedule = await _db.Schedules.FirstOrDefaultAsync(u => u.Id == id);
             if (schedule == null)
             {
                 return NotFound();
@@ -44,20 +48,18 @@ namespace SchedulerApi.Controllers
         [ProducesResponseType(StatusCodes.Status201Created)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
         [ProducesResponseType(StatusCodes.Status404NotFound)]
-        public ActionResult<ScheduleCreateDTO>CreateSchedule([FromBody]ScheduleCreateDTO createDTO) 
+        public async Task< ActionResult<ScheduleCreateDTO>>CreateSchedule([FromBody]ScheduleCreateDTO createDTO) 
         {
-            if(_db.Schedules.FirstOrDefault(u=>u.Name.ToLower()== createDTO.Name.ToLower()) != null)
+            if(await _db.Schedules.FirstOrDefaultAsync(u=>u.Name.ToLower()== createDTO.Name.ToLower()) != null)
             {
                 ModelState.AddModelError("CustomError", "Schedule already exixts");
                 return BadRequest(ModelState);
             }
-            Schedule schedule = new()
-            {
-                Name = createDTO.Name,
-                AppointmentDate = createDTO.AppointmentDate
-            };
-            _db.Schedules.Add(schedule);
-            _db.SaveChanges();
+            
+            Schedule schedule= _mapper.Map<Schedule>(createDTO);
+
+            await _db.Schedules.AddAsync(schedule);
+            await _db.SaveChangesAsync();
 
             return CreatedAtRoute("GetSchedule", new { id = schedule.Id }, schedule);
         }
@@ -65,35 +67,37 @@ namespace SchedulerApi.Controllers
         [HttpDelete("{id:int}",Name ="DeleteSchedule")]
         [ProducesResponseType(StatusCodes.Status204NoContent)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
-        public ActionResult DeleteSchedule(int id)
+        
+        public async Task< ActionResult> DeleteSchedule(int id)
         {
-            if(id== 0)
+            if(id== 0 )
             {
                 return BadRequest();
             }
-            var schedule = _db.Schedules.FirstOrDefault(u => u.Id == id);
+            if(await _db.Schedules.FirstOrDefaultAsync(u => u.Id != id) != null)
+            {
+                return BadRequest();
+            }
+            var schedule =await _db.Schedules.FirstOrDefaultAsync(u => u.Id == id);
             _db.Schedules.Remove(schedule);
-            _db.SaveChanges();
+            await _db.SaveChangesAsync();
             return NoContent();
         }
         [HttpPut("{id:int}",Name ="UpdateSchedule")]
         [ProducesResponseType(StatusCodes.Status204NoContent)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
         
-        public ActionResult<ScheduleUpdateDTO>UpdateSchedule(int id,[FromBody]ScheduleUpdateDTO updateDTO)
+        public async Task< ActionResult<ScheduleUpdateDTO>>UpdateSchedule(int id,[FromBody]ScheduleUpdateDTO updateDTO)
         {
             if(id!=updateDTO.Id)
             {
                 return BadRequest();
             }
-            Schedule schedule = new Schedule()
-            {
-                Id = updateDTO.Id,
-                Name = updateDTO.Name,
-                AppointmentDate = updateDTO.AppointmentDate,
-            };
+            
+            Schedule schedule=_mapper.Map<Schedule>(updateDTO);
+
             _db.Schedules.Update(schedule);
-            _db.SaveChanges();
+            await _db.SaveChangesAsync();
            return NoContent();
         }
     }
